Sync from coreutils.
* modules/lstat (Depends-on): Don't depend on xalloc.
(License): Change from GPL to LGPL, since this is now simply a
replacement for a libc function.

Eliminate the unwelcome (albeit unlikely) possibility of xmalloc
failure on deficient systems, and simplify gnulib lgpl dependencies.
* lib/lstat.c (rpl_lstat): Rewrite to use stat() in place of the
xmalloc/lstat combination.  Based on a patch from Bruno Haible.

* lib/xalloc-die.c: Remove unused definition of N_.

* m4/ls-mntd-fs.m4 (AC_FUNC_GETMNTENT): Invoke AC_CHECK_FUNCS(getmntent)
unconditionally so that tests of $ac_cv_func_getmntent (e.g., in
gl_LIST_MOUNTED_FILE_SYSTEMS) need not double-quote uses of that
variable, to accommodate the rare case in which getmntent is
available in none of the libraries checked.  This happens at
least on FreeBSD 5.0.