version-etc: match standards.texi style
This assumes that PACKAGE_BUGREPORTS and PACKAGE_PACKAGER_BUG_REPORTS
are generally email addresses, not URLs.

Coreutils switched to a similar style on 2009-02-01, with no complaints
(the difference is that coreutils inserts the program name, derived
from argv[0], into the bug reporting line).  Around that time,
standards.texi was also updated to mention this style.

* lib/version-etc.c (emit_bug_reporting_address): Drop periods,
and use <> only for URLs.

Signed-off-by: Eric Blake <ebb9@byu.net>