maint.mk: ignore multi-line copyright in NEWS
The old algorithm did not work for projects like m4 that used
long-hand copyright years that wrapped lines.  Meanwhile, this
change has no impact to single-line copyrights, like coreutils.

Changing the rest of the footer, such as bumping from GFDL 1.2
to GFDL 1.3, will still require the developer to rerun
'make update-NEWS-hash' to get the hash back in sync, but that
is probably less frequent.

* top/maint.mk (NEWS_hash): Add immunity to multi-line copyright.

Signed-off-by: Eric Blake <ebb9@byu.net>