maint.mk: detect incorrect GFDL usage
Similar to the existing sc_GPL_version.  Fortunately, the two licenses
have enough distinction in wording so that neither regexp is a false
positive for the other license.  This won't catch files that line wrap
differently than expected; as it was, the GFDL needed two regexp to
handle the use of GFDL in the m4 project.

* top/maint.mk (_GFDL_regexp, sc_GFDL_version): New rule.

Signed-off-by: Eric Blake <ebb9@byu.net>