maint.mk: check for [ as well as test
A word on the regex: \<\[ never matches, so the match instead
looks for a leading space in that case.  Although configure.ac
should generally use test instead of [ (because the latter has
to be spelled [[ because of m4 processing), libvirt had a case
of the latter that would have been detected with this rule.

* top/maint.mk (sc_prohibit_test_minus_ao): Extend test.
Based on a libvirt report by Matthias Bolte.

Signed-off-by: Eric Blake <eblake@redhat.com>