bootstrap: use "git -h" instead of "git --help"
Use "git -h" for testing for supported options instead of "git --help".
The short-form option only shows a summary, and doesn't layout the full
man page, which is much faster and saves a dependency.

(We found out that git uses man in a build environment of
 http://hydra.nixos.org/, which by default doesn't have man installed.)

Grep for the full option name in the summary, too.

Signed-off-by: Andreas Gruenbacher <agruen@suse.de>