New module `libposix'.
* modules/libposix: New module to facilitate an installable
library of posix modules.
(Makefile.am): Use lib_LTLIBRARIES to mark libposix.la as
installable.
(Depends-On): List all the modules found by `posix-modules',
except `strdup' which is spurious, and with the addition of
`alloca' to satisfy an otherwise undefined LTALLOCA definition;
and `progname' to provide a non-extern program_name symbol so that
compilation with `-no-undefined' symbols can work.